requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.28.0
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
pyyaml>=6.0.1
matplotlib>=3.8.0
tqdm>=4.66.0
Pillow>=10.0.0
pydicom>=2.4.0
